Unifying card model data set
